Reworked namespaces
