+++ src/Periturf.Tests/Verify/Evaluators/Logical/NotConditionEvaluatorTests.cs
@@ -16,22 +16,20 @@
 using FakeItEasy;
 using NUnit.Framework;
 using Periturf.Verify;
-using Periturf.Verify.Evaluators;
+using Periturf.Verify.Evaluators.Logical;
 using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Periturf.Tests.Verify.Evaluators
+namespace Periturf.Tests.Verify.Evaluators.Logical
 {
     [TestFixture]
-    class AndConditionEvaluatorTests
+    class NotConditionEvaluatorTests
     {
-        [TestCase(true, true, true)]
-        [TestCase(false, true, false)]
-        [TestCase(true, false, false)]
-        [TestCase(false, false, false)]
-        public async Task Given_ChildConditions_When_Evaluate_Then_ExpectedResult(bool conditionResult, bool condition2Result, bool parentConditionResult)
+        [TestCase(true, false)]
+        [TestCase(false, true)]
+        public async Task Given_ChildConditions_When_Evaluate_Then_ExpectedResult(bool conditionResult, bool parentConditionResult)
         {
             // Arrange
             var id = Guid.NewGuid();
@@ -39,18 +37,13 @@ namespace Periturf.Tests.Verify.Evaluators
             var evaluator = A.Dummy<IConditionEvaluator>();
             A.CallTo(() => evaluator.EvaluateAsync(A<CancellationToken>._)).Returns(conditionResult);
 
-            var evaluator2 = A.Dummy<IConditionEvaluator>();
-            A.CallTo(() => evaluator2.EvaluateAsync(A<CancellationToken>._)).Returns(condition2Result);
-
-            var spec = new AndConditionEvaluator(new List<IConditionEvaluator> { evaluator, evaluator2 });
+            var spec = new NotConditionEvaluator(evaluator);
 
             // Act
             var parentResult = await spec.EvaluateAsync();
 
             // Assert
             Assert.AreEqual(parentConditionResult, parentResult);
-            A.CallTo(() => evaluator.EvaluateAsync(A<CancellationToken>._)).MustHaveHappened();
-            A.CallTo(() => evaluator2.EvaluateAsync(A<CancellationToken>._)).MustHaveHappened();
         }
     }
 }
