+++ src/Periturf/Verify/Evaluators/Logical/AndConditionEvaluator.cs
@@ -13,17 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Periturf.Verify.Evaluators
+namespace Periturf.Verify.Evaluators.Logical
 {
     class AndConditionEvaluator : IConditionEvaluator
     {
-        private List<IConditionEvaluator> _evaluators;
+        private readonly List<IConditionEvaluator> _evaluators;
 
         public AndConditionEvaluator(IEnumerable<IConditionEvaluator> evaluators)
         {
